added authenticatin and authorization

+++ src/user/user.controller.ts
@@ -6,43 +6,57 @@ import {
   Param,
   Delete,
   UseGuards,
-  Request,
+  UnauthorizedException,
 } from '@nestjs/common';
 import { UserService } from './user.service';
 import { UpdateUserDto } from './dto/update-user.dto';
 import { LocalAuthGuard } from 'src/auth/guard/jwtAuth.guard';
+import { GetUser } from 'src/auth/decorator/get_user.decorator';
+import { User } from '@prisma/client';
 
+@UseGuards(LocalAuthGuard)
 @Controller('user')
 export class UserController {
   constructor(private readonly userService: UserService) {}
 
-  @UseGuards(LocalAuthGuard)
-  @Get('me')
-  getProfile(@Request() req) {
-    return req.user;
+  @Get('profile')
+  getProfile(@GetUser() user: User) {
+    return user;
   }
 
-  @UseGuards(LocalAuthGuard)
   @Get()
   findAll() {
     return this.userService.findAll();
   }
 
-  @UseGuards(LocalAuthGuard)
   @Get(':id')
   findOne(@Param('id') id: string) {
     return this.userService.findOne(+id);
   }
 
-  @UseGuards(LocalAuthGuard)
   @Patch(':id')
-  update(@Param('id') id: string, @Body() updateUserDto: UpdateUserDto) {
-    return this.userService.update(+id, updateUserDto);
+  update(
+    @GetUser() user: User,
+    @Param('id') id: string,
+    @Body() updateUserDto: UpdateUserDto,
+  ) {
+    if (user.id === +id) {
+      return this.userService.update(+id, updateUserDto);
+    } else {
+      throw new UnauthorizedException(
+        'You are not allowed to update other users',
+      );
+    }
   }
 
-  @UseGuards(LocalAuthGuard)
   @Delete(':id')
-  remove(@Param('id') id: string) {
-    return this.userService.remove(+id);
+  remove(@GetUser() user: User, @Param('id') id: string) {
+    if (user.id === +id) {
+      return this.userService.remove(+id);
+    } else {
+      throw new UnauthorizedException(
+        'You are not allowed to delete other users',
+      );
+    }
   }
 }
